+++ app/Http/Controllers/MoviesController.php
@@ -16,6 +16,8 @@ use app\Repositories\TagRepository;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Session;
+use Kurt\Imgur\Imgur;
+
 
 class MoviesController extends Controller
 {
@@ -24,18 +26,24 @@ class MoviesController extends Controller
     protected $actressRepository;
     protected $tagRepository;
 
+    protected $imgurService;
+    protected $imgurAllow;
+
     protected $indexOrder = ['order' => ['col' => 'updated_at',
         'dir' => 'desc']];
 
     public function __construct(MovieRepository $movieRepo,
                                 StudioRepository $studioRepo,
                                 ActressRepository $actressRepo,
-                                TagRepository $tagRepo)
+                                TagRepository $tagRepo,
+                                Imgur $imgur)
     {
         $this->movieRepository = $movieRepo;
         $this->studioRepository = $studioRepo;
         $this->actressRepository = $actressRepo;
         $this->tagRepository = $tagRepo;
+        $this->imgurService = $imgur;
+        $this->imgurAllow = config('setting.imgurService', false);
     }
 
     public function index(Request $request){
@@ -86,12 +94,24 @@ class MoviesController extends Controller
         try{
             // save thumbnail
             if(isset($data['thumbnail'])){
-                $data['thumbnail'] = storeImage($request->file('thumbnail'), $data['code'], 'custom.thumbnail_movie_path');
+                if ($this->imgurAllow){
+                    $imageModel = $this->imgurService->upload($request->file('thumbnail'));
+                    $data['thumbnail'] = $imageModel->getLink();
+                }
+                else {
+                    $data['thumbnail'] = storeImage($request->file('thumbnail'), $data['code'], 'custom.thumbnail_movie_path');
+                }
             }
 
             // save big image
             if(isset($data['image'])){
-                $data['image'] = storeImage($request->file('image'), $data['code'], 'custom.image_movie_path');
+                if ($this->imgurAllow){
+                    $imageModel = $this->imgurService->upload($request->file('image'));
+                    $data['image'] = $imageModel->getLink();
+                }
+                else {
+                    $data['image'] = storeImage($request->file('image'), $data['code'], 'custom.image_movie_path');
+                }
             }
 
             // check stored field
@@ -146,26 +166,38 @@ class MoviesController extends Controller
 
             // check new thumbnail
             if(isset($data['thumbnail'])){
-                // remove old thumbnail
-                if($movie->thumbnail != ''){
-                    $storagePath = storage_path(config('custom.thumbnail_movie_path') . $movie->thumbnail);
-                    File::delete($storagePath);
+                if ($this->imgurAllow){
+                    $imageModel = $this->imgurService->upload($request->file('thumbnail'));
+                    $data['thumbnail'] = $imageModel->getLink();
                 }
+                else {
+                    // remove old thumbnail
+                    if ($movie->thumbnail != '') {
+                        $storagePath = storage_path(config('custom.thumbnail_movie_path') . $movie->thumbnail);
+                        File::delete($storagePath);
+                    }
 
-                // save thumbnail
-                $data['thumbnail'] = storeImage($request->file('thumbnail'), $data['code'], 'custom.thumbnail_movie_path');
+                    // save thumbnail
+                    $data['thumbnail'] = storeImage($request->file('thumbnail'), $data['code'], 'custom.thumbnail_movie_path');
+                }
             }
 
             // check new image
             if(isset($data['image'])){
-                // remove old image
-                if($movie->image != ''){
-                    $storagePath = storage_path(config('custom.image_movie_path') . $movie->image);
-                    File::delete($storagePath);
+                if ($this->imgurAllow){
+                    $imageModel = $this->imgurService->upload($request->file('image'));
+                    $data['image'] = $imageModel->getLink();
                 }
+                else {
+                    // remove old image
+                    if ($movie->image != '') {
+                        $storagePath = storage_path(config('custom.image_movie_path') . $movie->image);
+                        File::delete($storagePath);
+                    }
 
-                // save big image
-                $data['image'] = storeImage($request->file('image'), $data['code'], 'custom.image_movie_path');
+                    // save big image
+                    $data['image'] = storeImage($request->file('image'), $data['code'], 'custom.image_movie_path');
+                }
             }
 
             // check stored field
